Move compose metrics parameters to Android Compose script
